+++ packages/moengage_cards/moengage_cards_android/CHANGELOG.md
@@ -4,9 +4,14 @@
 
 ## Release Version
 - Android
-  - Added support for AGP version `8.7.3` and above
-  - Compiled SDK Version updated to 35
-  - Min SDK Version updated to 23
+    - Compiled SDK Version updated to 35
+    - Min SDK Version updated to 23
+  
+# 25-03-2025
+
+## 4.2.0
+- Android
+    - Added support for AGP version `8.7.3` and above
 
 # 29-01-2025